Combining Structured Information into a Context for LLMs/Context_of_the_LLM/requirements.txt
gliner
stopwords
rank_bm25
google-generativeai
plotly
astropy
pandas
venn
requests
pandas
matplotlib
seaborn
nltk
numpy
Counter
stopwords
matplotlib_venn
dash  
dash-bootstrap-components
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0
scipy>=1.9.0
seaborn>=0.11.0
tqdm>=4.64.0
python-dateutil>=2.8.0
bitsandbytes>=0.39.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0